atomic source

+++ cosas/src/oscillator.cpp
@@ -18,7 +18,7 @@ int16_t BaseOscillator::next(const int32_t delta, const int32_t phi) {
    * the RelSource interface deals in delta samples - typically 1, but allowing
    * for more in case the output buffer underflows.  here we need to convert that
    * into absolute time for the AbsSource interfaces, which are all wavetables.
-   * since they are all wavetables they all loop at SAMPLE_RATE.  snice frequency
+   * since they are all wavetables they all loop at SAMPLE_RATE.  since frequency
    * is in SUBSAMPLE_BITS units you might naively think that we can calculate
    * time modulo SAMPLE_RATE * (SAMPLE_RATE << SUBSAMPLE_BITS).  actually a
    * factor of 0.5 lower because frequency cannot exceed SAMPLE_RATE / 2.
@@ -32,7 +32,7 @@ int16_t BaseOscillator::next(const int32_t delta, const int32_t phi) {
    */
   // increment time
   uint32_t frequency_val = LOAD(frequency);
-  tick += delta * frequency_val;
+  tick += delta * static_cast<int32_t>(frequency_val);
   if (tick > TIME_MODULUS) tick -= TIME_MODULUS;
   // convert phi to something like phase
   const int32_t phi_phase = (phi * frequency_val) >> PHI_FUDGE_BITS;  // arbitrary scaling
